Switch int to long types

+++ src/Main.java
@@ -1,7 +1,7 @@
 public class Main {
 
     public static void main(String[] args) {
-        SalesManager salesManager = new SalesManager(new int[]{1, 2, 6});
+        SalesManager salesManager = new SalesManager(new long[]{1, 2, 6});
 
         System.out.println(salesManager.max());
     }
